+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to
 ### Changed
 
 - 💄(frontend) improve comments highlights #1961
+- ♿️(frontend) improve BoxButton a11y and native button semantics
+  #2103
 
 ## [v4.8.3] - 2026-03-23